API call to TFL and sample data files.
A simple API written in the frontend in App.js.
Data is filtered and mapped to return only arrivals that are due in the
next 60 seconds, with selected data returned and sorted in order
of arrival time.
Data samples have been collected and stored in sampleData folder for
analysis and test use.