Data fetch repeats every x seconds.
Data is fetched every x seconds set by a variable at the top of the App.js file.
dataBlockDuration also controls the data that is filtered in the return to match
the data duration.
React strict mode is disbled to stop double logging of data.
Console log of tubeData is moved once more out of the data fetch function
to give the actual tubeData rather than sortedData to be 100% sure of the data
that we are working with.